+++ lib/feed2email/config.rb
@@ -7,40 +7,52 @@ module Feed2Email
     class InvalidConfigSyntaxError < StandardError; end
     class InvalidConfigDataTypeError < StandardError; end
     class MissingConfigOptionError < StandardError; end
+    class InvalidConfigOptionError < StandardError; end
+
+    SEND_METHODS = %w{file sendmail smtp}
 
     def initialize(path)
       @path = path
-      check
     end
 
     def [](option)
       config[option] # delegate
     end
 
-    def smtp_configured?
-      config['smtp_host'] && config['smtp_port'] && config['smtp_user'] &&
-        config['smtp_pass']
-    end
-
     private
 
-    def data
-      @data
-    end
-
-    def check
-      check_existence
-      check_permissions
-      check_syntax
-      check_data_type
-      check_recipient_existence
-      check_sender_existence
+    def check_data_type
+      if !data.is_a?(Hash)
+        raise InvalidConfigDataTypeError,
+          "Invalid data type (not a Hash) for config file #{path}"
+      end
     end
 
     def check_existence
       if !File.exist?(path)
         raise MissingConfigError, "Missing config file #{path}"
       end
+    end
+
+    def check_file
+      check_existence
+      check_permissions
+      check_syntax
+      check_data_type
+    end
+
+    def check_option(option)
+      if config[option].nil?
+        raise MissingConfigOptionError,
+          "Option #{option} missing from config file #{path}"
+      end
+    end
+
+    def check_options
+      check_recipient
+      check_sender
+      check_send_method
+      check_smtp_options if config['send_method'] == 'smtp'
     end
 
     def check_permissions
@@ -50,42 +62,47 @@ module Feed2Email
       end
     end
 
+    def check_recipient
+      check_option('recipient')
+    end
+
+    def check_send_method
+      unless SEND_METHODS.include?(config['send_method'])
+        raise InvalidConfigOptionError,
+          "Option send_method not one of: #{SEND_METHODS.join(' ')}"
+      end
+    end
+
+    def check_sender
+      check_option('sender')
+    end
+
+    def check_smtp_options
+      check_option('smtp_host')
+      check_option('smtp_port')
+      check_option('smtp_user')
+      check_option('smtp_pass')
+    end
+
     def check_syntax
       begin
-        load_yaml
+        data
       rescue Psych::SyntaxError
         raise InvalidConfigSyntaxError,
           "Invalid YAML syntax for config file #{path}"
       end
     end
 
-    def check_data_type
-      if !data.is_a?(Hash)
-        raise InvalidConfigDataTypeError,
-          "Invalid data type (not a Hash) for config file #{path}"
-      end
-    end
-
-    def check_recipient_existence
-      check_option_existence('recipient')
-    end
-
-    def check_sender_existence
-      check_option_existence('sender')
-    end
-
-    def load_yaml
-      @data = YAML.load(read_file)
-    end
-
-    def path; @path end
-
-    def read_file
-      File.read(path)
-    end
-
     def config
-      @config ||= defaults.merge(data)
+      return @config if @config
+      check_file
+      @config = defaults.merge(data)
+      check_options
+      @config
+    end
+
+    def data
+      @data ||= YAML.load(File.read(path))
     end
 
     def defaults
@@ -94,19 +111,16 @@ module Feed2Email
         'log_path'       => true,
         'log_shift_age'  => 0,
         'log_shift_size' => 1, # megabyte
+        'mail_path'      => File.join(ENV['HOME'], 'Mail'),
         'max_entries'    => 20,
         'send_delay'     => 10,
+        'send_method'    => 'file',
         'sendmail_path'  => '/usr/sbin/sendmail',
         'smtp_auth'      => 'login',
         'smtp_starttls'  => true,
       }
     end
 
-    def check_option_existence(option)
-      if data[option].nil?
-        raise MissingConfigOptionError,
-          "Option #{option} missing from config file #{path}"
-      end
-    end
+    def path; @path end
   end
 end